backend/requirements.txt
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
google-generativeai
email-validator
requests
motor
json-repair
fastapi
uvicorn
dnspython
